Added More API Endpoints

+++ src/main/java/dev/niranjan/BookMyShow/Repository/ShowSeatRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ShowSeatRepo extends JpaRepository<ShowSeat, Long> {
+public interface ShowSeatRepo extends JpaRepository<ShowSeat, Integer> {
 }
